% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COLOR3B.R
\name{COLOR3B}
\alias{COLOR3B}
\title{A color function}
\usage{
COLOR3B()
}
\value{
No particular return value
}
\description{
This function returns a set of colors that can be used for co-plotting
}
\examples{

barplot(1:3, names.arg=c('a','b','c'),ylab="some y",xlab="some x", col=COLOR3B())

}
\keyword{color}
\keyword{plot}

