% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COLOR4.R
\name{COLOR4}
\alias{COLOR4}
\title{A color function}
\usage{
COLOR4()
}
\value{
No particular return value
}
\description{
This function returns a set of colors that can be used for co-plotting
}
\examples{

barplot(1:4, names.arg=c('a','b','c','d'),ylab="some y",xlab="some x", col=COLOR4())

preset3a = get_preset(3)
pngpdf_preset(
  list(
      function(){barplot(1:4, names.arg=c('a','b','c','d'),ylab="some y",xlab="some x", col=COLOR4())},
      function(){plot(1:100,1:100,pch=19, type='o', xlab="")},
      function(){plot(rnorm(100),type='h')}
  ),
  'test5', stylefunc_blue, preset3a
)

}
\keyword{color}
\keyword{plot}

