% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COLOR5.R
\name{COLOR5}
\alias{COLOR5}
\title{A color function}
\usage{
COLOR5()
}
\value{
No particular return value
}
\description{
This function returns a set of colors that can be used for co-plotting
}
\examples{

barplot(1:5, names.arg=c('a','b','c','d','e'),ylab="some y",xlab="some x", col=COLOR5())

}
\keyword{color}
\keyword{plot}

