% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumhist.R
\name{cumhist}
\alias{cumhist}
\title{A plot function}
\usage{
cumhist(x, plot = TRUE, type = "l", ylab = "Cumulative fraction", ...)
}
\arguments{
\item{x}{a numeric vector.}

\item{plot}{If this value is true, a plot will be generated. If false, it will return a data frame in which x is the values in the input and cdf is the cumulative fraction.}

\item{type}{type of the plot. It is highly recommended not to use 'p','b' or 'o' if your data is very large. Defaults to 'l'.}

\item{ylab}{ylab. Defaults to 'Cumulative fraction".}

\item{...}{Additional plotting parameters.}
}
\value{
if plot=FALSE, then a data frame is returned. (see argument: plot)
}
\description{
This function either plots unsmoothed cumulative histogram or returns a data frame that can be used to plot such a cumulative histogram.
}
\examples{

# example 1
cumhist(rnorm(1000))

# example 2
x = cumhist(rnorm(1000),plot=F)
y = cumhist(runif(1000,-2,2),plot=F)
plot(x,type="l",col=1)
lines(y,col=2)

}
\seealso{
ecdf (stats package)
}
\keyword{histogram}
\keyword{plot}

