% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_axis.R
\name{exp_axis}
\alias{exp_axis}
\title{A plot axis wrapper function}
\usage{
exp_axis(x, axis_ind, n = 5)
}
\arguments{
\item{x}{Vector of numbers to plot or range of the numbers}

\item{axis_ind}{Axis number (1: bottom, 2: left, 3: top, 4: right)}

\item{n}{(Rough) number of tick marks that is used to find pretty tick mark positions. (parameter for pretty())}
}
\description{
This function generates an axis with a label in an exponential form, given a log-scale range.
}
\examples{

 logx=rnorm(100)
 logy=rnorm(100) + logx
 plot(logx, logy, xlab="x", ylab="y", axes=FALSE)
 exp_axis(range(logx),1)
 exp_axis(range(logy),2)
 box()
}
\keyword{axis}
\keyword{plot}

