% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_preset.R
\name{get_preset}
\alias{get_preset}
\title{Function that returns a 'preset', a list containing a preset layout of a plot.}
\usage{
get_preset(n = 1, layout = "")
}
\arguments{
\item{n}{Number of plots in the figure. (default 1) Currently, only n=1-3 are available.}

\item{layout}{The layout number. (default ''). Currently for n=2, 'h50' (horizontal division) and 'v50' (vertical division) are available. For n=1 and n=3, only one preset is available.}
}
\value{
a preset, list of plt vectors, width and height.
}
\description{
This function returns a preset object that contains a preset layout of a plot, which can be fed to the function pngpdf_preset(), to use a preselected layout for various arrangements of plots in the device.
}
\examples{

preset3a = get_preset(3)
pngpdf_preset( 
  list( 
      function(){barplot(1:4, names.arg=c('a','b','c','d'),ylab="some y",xlab="some x", col=COLOR4())},
      function(){plot(1:100,1:100,pch=19, type='o', xlab="")},
      function(){plot(rnorm(100),type='h')}
  ),
  'test5', stylefunc_blue, preset3a
)

}
\keyword{layout}
\keyword{plot}

