% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/plot_rainbowlines.R
\name{plot_rainbowlines}
\alias{plot_rainbowlines}
\title{A plot function}
\usage{
plot_rainbowlines(mat.list, samplenames = NULL, color = NULL, lty = 1,
  lwd = 1, xlab = "x", ylab = "y", main = "", cex.legend = 1,
  template1.args = NULL, add.func = NULL, ...)
}
\arguments{
\item{mat.list}{a list of two-column matrices or two-column data frames.}

\item{samplenames}{a vector of names that will be used as the legend.text. The length of samplenames must be the same. If not specified, the names of mat.list will be used. Defaults to NULL.}

\item{color}{a vector of colors that will be used for the plots. The length of color must be the same as the length of mat.list. If not specified, rainbow colors will be used automatically. Defaults to NULL.}

\item{lty}{line type.}

\item{lwd}{line width.}

\item{xlab}{x-axis label. Defaults to 'x'.}

\item{ylab}{y-axis label. Defaults to 'y'.}

\item{main}{title. Defaults to ''.}

\item{cex.legend}{the cex of the legend. Reduce this if you run out of vertical space for legend. Defaults to 1.}

\item{add.func}{an additional function to add to the plot (eg. abline).}

\item{...}{other graphical parameters.}

\item{template1.arg}{list of parameters for template1(), such as legend.plt.x.}
}
\value{
a vector of colors used.
}
\description{
This function generates a single plot with multiple lines of rainbow colors, given a list of matrices(or data frames).
}
\examples{
## example of three plots

par(mfrow=c(3,1))

# plot 1
mat.list = list("sd=1"=cumhist(rnorm(1000),plot=F),"sd=2"=cumhist(rnorm(2000,sd=2),plot=F))
plot_rainbowlines(mat.list)

# plot 2
plot_rainbowlines(mat.list,c("sd=1","sd=2"),color=c("violet","dodgerblue"),ylab="Cumulative fraction",main="Gaussian",lwd=2)

# plot 3
mat.list = lapply(1:10, function(sd) cumhist(rnorm(10000,sd=sd),plot=F) )
legend.text = paste("sd=",1:10,sep="")
plot_rainbowlines(mat.list,legend.text,x="values",ylab="Cumulative fraction",main="Gaussian",template1.args=list(legend.plt.x=0.8),add.func=function(){abline(h=0.5,lty=2)})
}
\keyword{lines}
\keyword{plot}
\keyword{rainbow}

