% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pngpdf.nodate.R
\name{pngpdf.nodate}
\alias{pngpdf.nodate}
\title{A wrapper function for pngpdf().}
\usage{
pngpdf.nodate(f, ...)
}
\arguments{
\item{...}{Other parameters to pass on to pngpdf.}

\item{plot.func}{Any plotting function.}
}
\value{
the returned value from plot.func will be returned.
}
\description{
This function generates both png and pdf files for a given plot function, but not adding the date in the output file name.
}
\examples{

pngpdf.nodate(function(){
  plot(1:10,1:10,type="o",pch=19,col=3)
},"myplot")

}
\keyword{pdf}
\keyword{plot}
\keyword{png}

