% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pngpdf_preset.R
\name{pngpdf_preset}
\alias{pngpdf_preset}
\title{Function that wraps pngpdf using a 'preset', to create a figure with a specific predefined layout.}
\usage{
pngpdf_preset(plotfuncs, name, stylefunc = stylefunc0,
  preset = get_preset(length(plotfuncs)), ...)
}
\arguments{
\item{plotfuncs}{List of plot functions.}

\item{name}{The prefix of the output png or pdf file.}

\item{stylefunc}{A style function that sets up graphic parameters.}

\item{preset}{A preset object with the same number of plts as the number of plot functions.}

\item{...}{Other parameters to be passed on to pngpdf()}
}
\value{
No particular return value
}
\description{
This function takes a list of plot functions that will be passed on to pngpdf, and draws each plot within its corresponding plt region specified by a preset object.
}
\examples{

preset3a = get_preset(3)
pngpdf_preset( 
  list( 
      function(){barplot(1:4, names.arg=c('a','b','c','d'),ylab="some y",xlab="some x", col=COLOR4())},
      function(){plot(1:100,1:100,pch=19, type='o', xlab="")},
      function(){plot(rnorm(100),type='h')}
  ),
  'test5', stylefunc_blue, preset3a
)


preset1a = get_preset(1)
pngpdf_preset( 
  list( 
      function(){ barplot(1:5, names.arg=c('a','b','c','d','e'),ylab="some y",xlab="some x", col=COLOR5()) }
   ),
   'test6', stylefunc0, preset1a
)

}
\keyword{layout}
\keyword{plot}

