% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stylefunc0.R
\name{stylefunc0}
\alias{stylefunc0}
\title{Function that sets a default style for plots}
\usage{
stylefunc0()
}
\value{
No particular return value
}
\description{
This function sets par values that can be used for a default style plots.
}
\examples{

stylefunc0()
hist(rnorm(100))


preset1a = get_preset(1)
pngpdf_preset( 
  list( 
      function(){ barplot(1:5, names.arg=c('a','b','c','d','e'),ylab="some y",xlab="some x", col=COLOR5()) }
   ),
   'test6', stylefunc0, preset1a
)

}
\keyword{par}
\keyword{plot}

