% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stylefunc_blue.R
\name{stylefunc_blue}
\alias{stylefunc_blue}
\title{Function that sets a style for plots with a blue background}
\usage{
stylefunc_blue()
}
\value{
No particular return value
}
\description{
This function sets par values that can be used for a particular style with a blue background.
}
\examples{

stylefunc_blue()
hist(rnorm(100))


preset3a = get_preset(3)
pngpdf_preset(
  list(
      function(){barplot(1:4, names.arg=c('a','b','c','d'),ylab="some y",xlab="some x", col=COLOR4())},
      function(){plot(1:100,1:100,pch=19, type='o', xlab="")},
      function(){plot(rnorm(100),type='h')}
  ),
  'test5', stylefunc_blue, preset3a
)

}
\keyword{blue}
\keyword{par}
\keyword{plot}

