% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/vioplot_list.R
\name{vioplot_list}
\alias{vioplot_list}
\title{A plot function}
\usage{
vioplot_list(x, horizontal = FALSE, axes = TRUE, Range = range(x, na.rm =
  T), add.boxplot = FALSE, drawRect = TRUE, col.vio = "light blue",
  col.box = 1, xlab = "", ylab = "", main = "")
}
\arguments{
\item{x}{a list of numeric vectors. If the list has names, the names will be used as labels. Otherwise 1:length(x) will be used.}

\item{horizontal}{if this is true, each violin is drawn horizontally, stacked vertically. If false, the other way around. Defaults to FALSE.}

\item{axes}{if this is true, axes will be drawn. Defaults to TRUE.}

\item{Range}{the range of values. This will be ylim if horizontal is false, xlim if horizontal is true. Defaults to the range of x.}

\item{add.boxplot}{adds boxplots on top of the violin plots if true. Defaults to FALSE.}

\item{drawRect}{adds thin boxplot-like rectangles on top of the violin plots if true. Defaults to TRUE.}

\item{col.vio}{color of violin plots. Defaults to 'light blue'.}

\item{col.box}{color of the boxplots. Defaults to 1 (black).}

\item{xlab}{x-axis label. Defaults to ''.}

\item{ylab}{y-axis label. Defaults to ''.}

\item{main}{title of the plot. Defaults to ''.}
}
\description{
This function generates a series of violins in one plot, given a list of numeric vectors.
}
\details{
This function requires package vioplot.
}
\examples{
# list without names
x=list(rnorm(1000),rnorm(1000,2,3),rnorm(1000,2,1))

# list with names
y=list(dist1=rnorm(1000),dist2=rnorm(1000,2,3),dist3=rnorm(1000,2,1))

par(mfrow=c(4,1))
vioplot_list(x,ylab="value")
vioplot_list(y,ylab="value",add.boxplot=T,col.vio="green3", drawRect=F)
vioplot_list(y,ylab="value",col.vio=1,drawRect=F)
vioplot_list(y,xlab="value",col.vio=1,drawRect=F,horizontal=T)
}
\keyword{boxplot}
\keyword{plot}
\keyword{violinplot}

